+++ src/app/app/admin/categories/page.tsx
