+++ src/app/app/categories/[slug]/page.tsx
