+++ src/app/app/courses/[slug]/page.tsx
