--- src/app/app/courses/search/page.tsx
@@ -1,3 +1,0 @@
-export default function CourseSearchPage() {
-    return (<>This page is blank</>);
-}
