+++ src/app/app/instruct/courses/[id]/curriculum/page.tsx
